# web/openapi.json
{
 "components": {
  "schemas": {
   "EntityRef": {
    "description": "实体引用：ID 或名称+类型，不信任用户输入的自由表名/列名。",
    "properties": {
     "entity_id": {
      "anyOf": [
       {
        "format": "uuid",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Entity Id"
     },
     "entity_type": {
      "pattern": "^(Company|Security|Product|Theme|Concept|Industry)$",
      "title": "Entity Type",
      "type": "string"
     },
     "name_hint": {
      "anyOf": [
       {
        "maxLength": 200,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name Hint"
     }
    },
    "required": [
     "entity_type"
    ],
    "title": "EntityRef",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
   "NumericFilter": {
    "description": "财务筛选：只使用 financial_metric_catalog 白名单中的指标。\n\nPeriodRule 与 operator/fiscal_years 必须一致（验收 5 的口径约束）：\n- LAST_N_FY            ⇒ fiscal_years == N；\n- LAST_3_FY_TOTAL_POSITIVE ⇒ operator=TOTAL_POSITIVE 且 fiscal_years=3；\n- CONSECUTIVE_3_FY_POSITIVE ⇒ operator=CONSECUTIVE_POSITIVE 且 fiscal_years=3。",
    "properties": {
     "fiscal_years": {
      "default": 3,
      "maximum": 10.0,
      "minimum": 1.0,
      "title": "Fiscal Years",
      "type": "integer"
     },
     "metric_code": {
      "maxLength": 64,
      "minLength": 2,
      "title": "Metric Code",
      "type": "string"
     },
     "operator": {
      "$ref": "#/components/schemas/PathOperator"
     },
     "period_rule": {
      "default": "LAST_3_FY",
      "maxLength": 50,
      "title": "Period Rule",
      "type": "string"
     },
     "threshold": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Threshold"
     }
    },
    "required": [
     "metric_code",
     "operator"
    ],
    "title": "NumericFilter",
    "type": "object"
   },
   "PathOperator": {
    "enum": [
     "TOTAL_POSITIVE",
     "CONSECUTIVE_POSITIVE",
     "MIN_VALUE",
     "MAX_VALUE"
    ],
    "title": "PathOperator",
    "type": "string"
   },
   "QueryIntent": {
    "enum": [
     "ENTITY_LOOKUP",
     "SEMANTIC_SCREEN",
     "EXPLAIN_RELATION",
     "TIMELINE",
     "COMPARE",
     "EVIDENCE_SEARCH"
    ],
    "title": "QueryIntent",
    "type": "string"
   },
   "QueryPlan": {
    "description": "受控查询计划：Planner 输出的唯一合法形态。",
    "properties": {
     "as_of": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "As Of"
     },
     "evidence_required": {
      "default": true,
      "title": "Evidence Required",
      "type": "boolean"
     },
     "intent": {
      "$ref": "#/components/schemas/QueryIntent"
     },
     "known_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Known At"
     },
     "max_results": {
      "default": 50,
      "maximum": 100.0,
      "minimum": 1.0,
      "title": "Max Results",
      "type": "integer"
     },
     "numeric_filter": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/NumericFilter"
       },
       {
        "type": "null"
       }
      ]
     },
     "object_entity": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/EntityRef"
       },
       {
        "type": "null"
       }
      ]
     },
     "plan_id": {
      "format": "uuid",
      "title": "Plan Id",
      "type": "string"
     },
     "semantic_filter": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/SemanticFilter"
       },
       {
        "type": "null"
       }
      ]
     },
     "subject": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/EntityRef"
       },
       {
        "type": "null"
       }
      ]
     },
     "timeout_seconds": {
      "default": 30,
      "maximum": 60.0,
      "minimum": 1.0,
      "title": "Timeout Seconds",
      "type": "integer"
     }
    },
    "required": [
     "intent"
    ],
    "title": "QueryPlan",
    "type": "object"
   },
   "QueryRequest": {
    "description": "查询请求：直接传入受控 QueryPlan（LLM Planner 后续可选）。",
    "properties": {
     "natural_question": {
      "anyOf": [
       {
        "maxLength": 2000,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Natural Question"
     },
     "plan": {
      "$ref": "#/components/schemas/QueryPlan"
     }
    },
    "required": [
     "plan"
    ],
    "title": "QueryRequest",
    "type": "object"
   },
   "QueryResponse": {
    "description": "Grounded 结构化响应：不添加结构化结果中不存在的任何实体。",
    "properties": {
     "conflicts": {
      "items": {
       "type": "string"
      },
      "title": "Conflicts",
      "type": "array"
     },
     "degradation_notes": {
      "items": {
       "type": "string"
      },
      "title": "Degradation Notes",
      "type": "array"
     },
     "degraded": {
      "default": false,
      "title": "Degraded",
      "type": "boolean"
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     },
     "excluded": {
      "items": {
       "type": "string"
      },
      "title": "Excluded",
      "type": "array"
     },
     "intent": {
      "$ref": "#/components/schemas/QueryIntent"
     },
     "period_rule": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Period Rule"
     },
     "query_id": {
      "format": "uuid",
      "title": "Query Id",
      "type": "string"
     },
     "results": {
      "items": {
       "$ref": "#/components/schemas/GroundedResult"
      },
      "title": "Results",
      "type": "array"
     },
     "status": {
      "$ref": "#/components/schemas/QueryStatus"
     },
     "trace_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Trace Id"
     },
     "unknowns": {
      "items": {
       "type": "string"
      },
      "title": "Unknowns",
      "type": "array"
     }
    },
    "required": [
     "query_id",
     "status",
     "intent"
    ],
    "title": "QueryResponse",
    "type": "object"
   },
   "QueryStatus": {
    "enum": [
     "RECEIVED",
     "PLANNED",
     "RUNNING",
     "SUCCEEDED",
     "FAILED",
     "CANCELLED",
     "REJECTED",
     "DEGRADED"
    ],
    "title": "QueryStatus",
    "type": "string"
   },
   "ScreenRequest": {
    "description": "筛选请求：语义条件 + 财务条件的简化入口。\n\n前置约束与 QueryPlan 模型一致（非法组合在请求层即 422）；\n注入载荷在计划构造层被拒（同样走受控 422）。",
    "properties": {
     "as_of": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "As Of"
     },
     "business_stage": {
      "anyOf": [
       {
        "maxLength": 50,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Business Stage"
     },
     "concept_name": {
      "anyOf": [
       {
        "maxLength": 200,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Concept Name"
     },
     "evidence_required": {
      "default": true,
      "title": "Evidence Required",
      "type": "boolean"
     },
     "fiscal_years": {
      "default": 3,
      "title": "Fiscal Years",
      "type": "integer"
     },
     "known_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Known At"
     },
     "max_results": {
      "default": 50,
      "maximum": 100.0,
      "minimum": 1.0,
      "title": "Max Results",
      "type": "integer"
     },
     "metric_code": {
      "anyOf": [
       {
        "maxLength": 64,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Metric Code"
     },
     "operator": {
      "$ref": "#/components/schemas/PathOperator",
      "default": "TOTAL_POSITIVE"
     },
     "period_rule": {
      "anyOf": [
       {
        "maxLength": 50,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Period Rule"
     },
     "threshold": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Threshold"
     }
    },
    "title": "ScreenRequest",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "ontologyMVP API",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/companies": {
   "get": {
    "description": "公司列表/搜索：名称前缀或包含匹配（参数化 LIKE，无自由 SQL）。\n\n每行携带证券代码与 ACCEPTED 量产 Claim 计数（列表页核心信息）；\n排序固定 canonical_name（确定性分页）。",
    "operationId": "list_companies_api_v1_companies_get",
    "parameters": [
     {
      "in": "query",
      "name": "q",
      "required": false,
      "schema": {
       "maxLength": 200,
       "title": "Q",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "stage",
      "required": false,
      "schema": {
       "maxLength": 50,
       "title": "Stage",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 50,
       "maximum": 100,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "offset",
      "required": false,
      "schema": {
       "default": 0,
       "maximum": 10000,
       "minimum": 0,
       "title": "Offset",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response List Companies Api V1 Companies Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Companies",
    "tags": [
     "query"
    ]
   }
  },
  "/api/v1/companies/{company_id}": {
   "get": {
    "operationId": "get_company_api_v1_companies__company_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "company_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Company Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Company Api V1 Companies  Company Id  Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Company",
    "tags": [
     "query"
    ]
   }
  },
  "/api/v1/companies/{company_id}/claims": {
   "get": {
    "operationId": "get_company_claims_api_v1_companies__company_id__claims_get",
    "parameters": [
     {
      "in": "path",
      "name": "company_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Company Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 50,
       "maximum": 200,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Company Claims Api V1 Companies  Company Id  Claims Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Company Claims",
    "tags": [
     "query"
    ]
   }
  },
  "/api/v1/companies/{company_id}/timeline": {
   "get": {
    "description": "主体时间线：ACCEPTED Claim 的双时态视图。",
    "operationId": "get_company_timeline_api_v1_companies__company_id__timeline_get",
    "parameters": [
     {
      "in": "path",
      "name": "company_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Company Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Company Timeline Api V1 Companies  Company Id  Timeline Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Company Timeline",
    "tags": [
     "query"
    ]
   }
  },
  "/api/v1/overview/recent-runs": {
   "get": {
    "description": "最近采集/标准化运行（首页\"最近活动\"卡片）。",
    "operationId": "overview_recent_runs_api_v1_overview_recent_runs_get",
    "parameters": [
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 10,
       "maximum": 50,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Overview Recent Runs Api V1 Overview Recent Runs Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Overview Recent Runs",
    "tags": [
     "query"
    ]
   }
  },
...
  "/api/v1/query": {
   "post": {
    "description": "执行受控查询计划。只读——不修改事实/财务/图（除查询审计）。",
    "operationId": "execute_query_api_v1_query_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/QueryRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/QueryResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Execute Query",
    "tags": [
     "query"
    ]
   }
  },
  "/api/v1/screen": {
   "post": {
    "description": "筛选：概念/阶段 + 财务条件的简化查询。\n\n计划构造经由工厂延迟到 _execute_plan 的受控 try 块内执行——\n非法组合与注入载荷得到 422，而非未捕获 500。",
    "operationId": "execute_screen_api_v1_screen_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ScreenRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/QueryResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Execute Screen",
    "tags": [
     "query"
    ]
   }
  },
...
 }
}